feat: Add initial project structure with configuration files and database setup scripts
- Created .gitignore to exclude target directory and mail properties.
- Updated pom.xml to change Java version to 11 and updated dependencies for Java EE.
- Added setup-database.sh for database initialization.
- Introduced start-server.sh to launch the Jetty server.
- Refactored clientDao to include password and date of birth in client model.
- Enhanced DBConnection to support environment variable configuration.
- Implemented EmailServlet for manual email notifications.
- Updated various servlets to improve error handling and user feedback.
- Added PDF generation for transaction reports with client details.
- Introduced utility functions for generating unique IDs in servlets.
- Improved data handling in fraisEnvoiDao and tauxDao.
- Added new methods for managing client and transaction data.
- Enhanced error messages and validation throughout the application.

+++ src/main/java/com/test/monprojetjsp/dao/fraisEnvoiDao.java
@@ -135,7 +135,7 @@ public class fraisEnvoiDao {
                 f.setIdfrais(rs.getString("idfrais"));
                 f.setMontant1(rs.getInt("montant1"));
                 f.setMontant2(rs.getInt("montant2"));
-                f.setFrais(rs.getInt("frais"));
+                f.setFrais(rs.getFloat("frais"));
             }
         }catch(Exception e){
             e.printStackTrace();
